OEL-2710: Move Person to JS.
